requirements.txt
Flask
PyMuPDF
pytesseract
Pillow
pandas
SQLAlchemy
